% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASA.R
\name{ASA}
\alias{ASA}
\title{Animism Scale for Adults (ASA)}
\usage{
ASA(
  label = "ASA",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f"
)
}
\arguments{
\item{label}{(character scalar) label given to the results section.}

\item{dict}{internationalization dictionary.}

\item{default_lang}{(scalar character) default language.}
}
\description{
This function defines an Animism Scale for Adults (ASA) module for
incorporation into a psychTestR timeline.
The ASA module uses two of the three original factors of the ASA
(池内 [Ikeuchi], 2010),
the 自然物の神格化 ("Deification of Natural Objects"), and the
所有者の分身化 ("Owner's Incarnation").
}
\references{
池内裕美. (2010). 成人のアニミズム的思考: 自発的喪失としてのモノ供養の心理.
\emph{社会心理学研究}, \emph{25}(3), 167--177.
https://doi.org/10.14966/jssp.KJ00006203282
[Ikeuchi, H. (2010). Animistic thinking in adults: The memorial service for
dolls as a voluntary loss. \emph{The Japanese Journal of Social Psychology},
\emph{25}(3), 167--177.]
}
