% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{audio_vas_page_battery}
\alias{audio_vas_page_battery}
\title{Battery of Audio Visual Analog Scale Pages}
\usage{
audio_vas_page_battery(
  battery_label,
  prompt,
  min_label,
  max_label,
  min_numeric,
  max_numeric,
  value,
  num_stimuli,
  stimulus_prefix_pattern = "s\%03d",
  base_url,
  randomise_at_runtime = FALSE,
  type = "mp3",
  wait = TRUE,
  loop = FALSE,
  btn_play_prompt = if (!show_controls) "Click here to play",
  show_controls = TRUE,
  allow_download = FALSE,
  autoplay = "autoplay",
  save_answer = TRUE,
  button_text = "Next",
  on_complete = NULL,
  admin_ui = NULL,
  step,
  hide_numeric_values = TRUE,
  fill_bar = TRUE,
  round = FALSE,
  ticks = FALSE,
  animate = FALSE,
  slider_width = "300px",
  sep = ",",
  pre = NULL,
  post = NULL
)
}
\arguments{
\item{battery_label}{(character scalar) Name for the results section of the
battery. Individual page labels (see \code{psychTestR::page()}) are generated
from \code{battery_label} and \code{stimulus_prefix_pattern}.}

\item{prompt}{Prompt to display (character scalar or Shiny tag object).}

\item{min_label}{(scalar character) Label for the left endpoint of the scale,
which corresponds to the minimum value of the slider.}

\item{max_label}{(scalar character) Label for the right endpoint of the
scale, which corresponds to the maximum value of the slider.}

\item{min_numeric}{(scalar numeric) Minimum value for the slider.}

\item{max_numeric}{(scalar numeric) Maximum value for the slider.}

\item{value}{(scalar numeric) Initial value of the slider.
Must lie between \code{min} and \code{max}.}

\item{num_stimuli}{(integer-like scalar) Number of stimuli / pages.}

\item{stimulus_prefix_pattern}{(character scalar) File name pattern of the
stimuli (internally used in \code{base::sprintf()}).
For example, \code{num_stimuli = 4}, \code{stimulus_prefix_pattern = "s\%03d"}, and
\code{type = "mp3"} results in the file names \code{s001.mp3}, \code{s002.mp3},
\code{s003.mp3}, and \code{s004.mp3}.}

\item{base_url}{(character scalar) URL without the stimulus file names.}

\item{type}{Audio type (e.g. 'mp3'). Defaults to the provided file extension.}

\item{wait}{Whether to wait for the audio to finish before displaying
the response buttons.}

\item{loop}{Whether the audio should loop.}

\item{btn_play_prompt}{Text to display as a prompt for starting the audio.
Ordinarily the participant will not see this,
but it might appear if the internet connection is poor,
or if the participant refreshes the page.
Only used if show_controls is FALSE.}

\item{show_controls}{Whether or not to show audio controls to the participant,
so that they can control audio playback.}

\item{allow_download}{Whether the participant is given a button to download
the audio file; only relevant if show_controls is TRUE.}

\item{autoplay}{Whether the audio should start playing immediately
default is "autoplay"}

\item{save_answer}{Whether or not to save the answer.}

\item{button_text}{Text for the submit button (character scalar).}

\item{on_complete}{Optional function to execute on leaving the page
(after successful validation).
The argument list should include \code{...},
and any of:
\code{state}, the participant's state object;
\code{answer}, the participant's most recent answer;
\code{input}, the current page's Shiny input object;
\code{session}, the current Shiny session object;
\code{opt}, the test's option list as created by \code{test_options()}.}

\item{admin_ui}{Optional UI component for the admin panel.}

\item{step}{Specifies the interval between each selectable value on the
slider. Either \code{NULL}, the default, which uses a heuristic to determine the
step size or a single number. If the values are dates, \code{step} is in days;
if the values are date-times, \code{step} is in seconds.}

\item{hide_numeric_values}{(logical scalar) Whether or not to hide the
numeric values in the UI, that is, \code{min_numeric}, \code{max_numeric}, and \code{value}.}

\item{fill_bar}{(logical scalar) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Whether or not to fill the bar from
the minimum to the position of the slider.
Currently only works for the first occurrence of a slider.}

\item{round}{\code{TRUE} to round all values to the nearest integer;
\code{FALSE} if no rounding is desired; or an integer to round to that
number of digits (for example, 1 will round to the nearest 10, and -2 will
round to the nearest .01). Any rounding will be applied after snapping to
the nearest step.}

\item{ticks}{\code{FALSE} to hide tick marks, \code{TRUE} to show them
according to some simple heuristics.}

\item{animate}{\code{TRUE} to show simple animation controls with default
settings; \code{FALSE} not to; or a custom settings list, such as those
created using \code{\link[shiny:animationOptions]{animationOptions()}}.}

\item{slider_width}{(scalar) Width of the slider. Must be valid for use as a
CSS unit of length.}

\item{sep}{Separator between thousands places in numbers.}

\item{pre}{A prefix string to put in front of the value.}

\item{post}{A suffix string to put after the value.}
}
\description{
Battery of Audio Visual Analog Scale Pages
}
