% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educational-degree.R
\name{education_page}
\alias{education_page}
\title{Educational degree}
\usage{
education_page(
  label = "educational_degree",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f"
)
}
\arguments{
\item{label}{(scalar character) Label for the current page.}

\item{dict}{internationalization dictionary.}

\item{default_lang}{(scalar character) default language.}
}
\description{
This function creates a page to ask for the highest educational degree.
The choices are oriented towards the ISCED categories.
}
