% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotional-baseline.R
\name{emotional_baseline_page}
\alias{emotional_baseline_page}
\title{Individual emotional baseline}
\usage{
emotional_baseline_page(
  label = "emotional_baseline",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f"
)
}
\arguments{
\item{label}{Label for the current page.}

\item{dict}{internationalization dictionary.}

\item{default_lang}{Default language for the page.}
}
\description{
This function creates a page examining the emotional baseline of
a participant.
}
