% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotional-baseline.R
\name{emotional_baseline_single_pages}
\alias{emotional_baseline_single_pages}
\title{Individual emotional baseline (multi page)}
\usage{
emotional_baseline_single_pages(
  label = "emotional_baseline",
  dict = vocaloidproject::vocaloidproject_dict,
  vertical = FALSE,
  default_lang = "de_f"
)
}
\arguments{
\item{label}{Label for the current page.}

\item{dict}{internationalization dictionary.}

\item{vertical}{(logical scalar) Whether or not to arrange the
response buttons vertically.}

\item{default_lang}{Default language for the page.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function creates multiple pages examining the emotional baseline of
a participant.
}
