% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp4_international1_stimuli}
\alias{exp4_international1_stimuli}
\title{Stimulus coding for International Experiment 1}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 120 rows and 2 columns.
}
\usage{
exp4_international1_stimuli
}
\description{
Stimuli for International Experiment 1 (also referred to as Experiment 4)
are 30 excerpts in 4 conditions.
The original file names followed the naming scheme \code{EXCERPT_CONDITION.mp3},
where \code{EXCERPT} is a number from 1 to 30 denoting the excerpt and
\code{CONDITION} is a string denoting the condition.
For use in the experiment file names were re-coded to \code{CEE.mp3}, where
\code{C} denotes the condition and \code{EE} is a width-two integer denoting the
excerpt (i.e., Excerpt 2 is represented by \code{02}).
}
\keyword{datasets}
