% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment4.R
\name{experiment4}
\alias{experiment4}
\title{Experiment 4 (international1)}
\usage{
experiment4(
  title = "",
  dict = vocaloidproject::vocaloidproject_dict,
  session_number = 1,
  debug = FALSE,
  technical_error_back_link = "",
  no_consent_back_link = "",
  unmet_requirements_back_link = "",
  year_range = c(1955, 2007),
  gms_subscales = c("General"),
  follow_up_link = "",
  admin_password = "vocaloid",
  researcher_email = "kilian.vogt@hmtm-hannover.de",
  researcher_email_language_list = NULL,
  languages = c("de_f", "ja"),
  logo = NULL,
  language_url_param = "l",
  language_url_codes = c(de = "ger", de_f = "ger", en = "eng", ja = "jpn")
)
}
\arguments{
\item{title}{The test's title.
This can either be an unnamed character scalar,
in which case the same title will be shown irrespective of internationalisation,
or a named vector of titles with the names
corresponding to language codes.}

\item{dict}{internationalization dictionary.}

\item{session_number}{(integer-like scalar) sets the session number.}

\item{debug}{(logical scalar) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} whether
or not to display debug information.}

\item{technical_error_back_link}{(character scalar) link for returning to
Prolific when URL parameter \code{uses} or \code{udes} is missing.
It has to contain the placeholder \verb{\%s} for \code{p_id} to redirect
participants correctly.}

\item{no_consent_back_link}{(character scalar) link for redirecting
participants who do not give consent.}

\item{unmet_requirements_back_link}{(character scalar) link for redirecting
participants when they do not meet study requirements.}

\item{year_range}{(length 2 int vector) The minimum and maximum year for birth date dropdown boxes.}

\item{gms_subscales}{(character vector) passed to the \code{subscales} argument of
\code{\link[psyquest:GMS]{psyquest::GMS()}}.}

\item{follow_up_link}{(character scalar) link for redirecting participants
to the next section of the experiment (on a different platform).
It has to contain 3 placeholders, \verb{\%s}, for \code{p_id}, \code{uses}, and \code{udes}.
The URL parameter for language, \code{language_url_param}, and its placeholder
are appended automatically.}

\item{admin_password}{Password to access the admin panel.}

\item{researcher_email}{Researcher's email; used in participant help message.}

\item{researcher_email_language_list}{(named character vector or \code{NULL})
provide language specific researcher emails as vector elements.
Names have to match the language codes.
If non-\code{NULL}, \code{researcher_email_language_list} takes precedence over
\code{researcher_email}.}

\item{languages}{Character vector of languages that may be selected
via the URL parameter 'language'. If no language is provided by
the URL parameter, defaults to the first language in this vector.
Languages should be encoded according to ISO 639-2 conventions, lower-case,
e.g. 'en'.}

\item{logo}{Path to a logo to display in the header (optional).}

\item{language_url_param}{(character scalar or \code{NULL}) URL parameter to pass
on the test language of a participant.
If non-\code{NULL}, the URL parameter will be appended to the \code{back_link} and
the language will be read from the participant's session info.}

\item{language_url_codes}{(named character vector or \code{NULL})
A named character vector will be used when \code{language_url_param} is
non-\code{NULL} to translate the psychTestR language codes
(in most cases two lower-case letters according to ISO 639-2 conventions)
to the codes for the redirect link
(e.g., SoSci Survey uses three letter codes like \code{ger} for German, \code{eng} for
English, or \code{jpn} for Japanese).
If \code{NULL}, the character scalar retrieved by
\code{\link[psychTestR:get_session_info]{psychTestR::get_session_info()}}\verb{$language} is used.}
}
\description{
psychTestR battery for Experiment 4 (International Experiment 1) in
3 sessions (see \code{\link[=international1]{international1()}} for a 2-session variant).
It is designed to redirect to SoSci Survey for the stimulus ratings and
some questionnaires that look much prettier there.
}
