% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{experiment4_designs}
\alias{experiment4_designs}
\title{Experiment 4 Stimulus Order}
\format{
An object of class \code{list} of length 4.
}
\usage{
experiment4_designs
}
\description{
This list contains the four stimulus presentation designs for Experiment 4.
}
\keyword{datasets}
