% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment4-monitor.R
\name{experiment4_monitor}
\alias{experiment4_monitor}
\title{Monitor App for Experiment 4}
\usage{
experiment4_monitor(
  battery_folder_1 = "",
  battery_folder_2 = "",
  battery_folder_3 = "",
  sosci_data = NULL,
  data_pw = "supersecretpassword"
)
}
\arguments{
\item{battery_folder_1}{(character scalar) This is the name of the folder
on your shiny server where the \code{app.R} containing
\code{experiment4(session_number = 1, ...)}
lives (i.e. app for Session 1).}

\item{battery_folder_2}{(character scalar) This is the name of the folder
on your shiny server where the \code{app.R} containing
\code{experiment4(session_number = 2, ...)}
lives (i.e. app for Session 2).}

\item{battery_folder_3}{(character scalar) This is the name of the folder
on your shiny server where the \code{app.R} containing
\code{experiment4(session_number = 3, ...)}
lives (i.e. app for Session 3).}

\item{sosci_data}{(\code{NULL} or character scalar) if non-\code{NULL}, a URL to
load data from (provided by SoSci Survey).}

\item{data_pw}{(character scalar) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
set a password for data access.
Caution: This is a very simple implementation.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calls a simple monitor app for Experiment 4 as defined by
\code{\link[=experiment4]{experiment4()}}.
The function assumed that it is running on the same shiny server as
\code{\link[=experiment4]{experiment4()}}.
}
