% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-language.R
\name{first_language_page}
\alias{first_language_page}
\title{First language}
\usage{
first_language_page(
  label = "first_language",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f"
)
}
\arguments{
\item{label}{(scalar character) Label for the current page.}

\item{dict}{internationalization dictionary.}

\item{default_lang}{(scalar character) default language.}
}
\description{
This function creates a page to ask for the first language.
For this research project, the choices are German (\code{de}), Japanese (\code{ja}),
English (\code{en}), and other language (\code{other}).
}
