% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ja-DEG.R
\name{ja_DEG}
\alias{ja_DEG}
\title{Japanese variant of DEG Age and Gender subscales}
\usage{
ja_DEG(
  label = "DEG",
  dict = vocaloidproject::vocaloidproject_dict,
  subscales = c(),
  languages = "ja",
  year_range = c(1930, 2013),
  ...
)
}
\arguments{
\item{label}{(Character scalar) Three uppercase letter acronym of the questionnaire.
This is also the label given to the results in the output file.}

\item{dict}{Dictionary used for internationalization.}

\item{subscales}{(Character vector) The subscales to be included in the questionnaire.
Possible subscales are "Best Shot", "Hearing Impairment", "Type of Hearing Impairment", "Gender", "Age", "Nationality", "Country Formative Years", "First Language", "Second Language", and "Handedness".
If no subscales are provided all subscales are selected.}

\item{year_range}{(length 2 int vector) The minimum and maximum year for birth date dropdown boxes.}

\item{...}{Further arguments to be passed to \code{\link[psyquest]{DEG}()} (e.g., can be used to set choice of country of residence.)}
}
\description{
\code{\link[psyquest:DEG]{psyquest::DEG()}} provides self-report pages for age and gender.
The \code{ja_DEG()} function provides an age page that asks for year and month
of birth (instead of month and year) better reflecting the Japanese date
format.
The gender self-report page is identical to the one in \code{\link[psyquest:DEG]{psyquest::DEG()}}.
Use \code{ja_DEG()} and \code{\link[psyquest:DEG]{psyquest::DEG()}} in conjunction with
\code{\link[psychTestR:conditional]{psychTestR::conditional()}} to display the adequate date format based on
the language.
}
