% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-language.R
\name{japanese_skills_page}
\alias{japanese_skills_page}
\title{Japanese skills}
\usage{
japanese_skills_page(
  label = "japanese_skills",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f"
)
}
\description{
This function creates a page to ask for Japanese skills.
}
