% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-language.R
\docType{data}
\name{language_codes}
\alias{language_codes}
\title{Language codes}
\usage{
language_codes
}
\description{
A named vector with two-letter language codes.
Names are the corresponding dictionary keys.
}
