% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ja-DEG.R
\name{make_ui_ja_year_and_month_select}
\alias{make_ui_ja_year_and_month_select}
\title{Make year and month selectboxes adapted to Japanese}
\usage{
make_ui_ja_year_and_month_select(
  id = "response_ui",
  min_year = 1930,
  max_year = 2013,
  show_month = TRUE
)
}
\arguments{
\item{id}{HTML ID for the div containing the selectboxes.}

\item{min_year}{minimum year to display in the year selectbox.}

\item{max_year}{maximum year to display in the year selectbox.}

\item{show_month}{(flag) whether or not to display the selectbox for month.}
}
\description{
Creates html code for year and month selectboxes.
}
