% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-elements.R
\name{make_ui_vas}
\alias{make_ui_vas}
\title{Create UI for Visual Analog Scale}
\usage{
make_ui_vas(
  min_label,
  max_label,
  min_numeric,
  max_numeric,
  value,
  step,
  hide_numeric_values = TRUE,
  round = FALSE,
  ticks = FALSE,
  animate = FALSE,
  slider_width = "300px",
  sep = ",",
  pre = NULL,
  post = NULL
)
}
\arguments{
\item{min_label}{(scalar character) Label for the left endpoint of the scale,
which corresponds to the minimum value of the slider.}

\item{max_label}{(scalar character) Label for the right endpoint of the
scale, which corresponds to the maximum value of the slider.}

\item{min_numeric}{(scalar numeric) Minimum value for the slider.}

\item{max_numeric}{(scalar numeric) Maximum value for the slider.}

\item{value}{(scalar numeric) Initial value of the slider.
Must lie between \code{min} and \code{max}.}

\item{step}{Specifies the interval between each selectable value on the
slider. Either \code{NULL}, the default, which uses a heuristic to determine the
step size or a single number. If the values are dates, \code{step} is in days;
if the values are date-times, \code{step} is in seconds.}

\item{hide_numeric_values}{(logical scalar) Whether or not to hide the
numeric values in the UI, that is, \code{min_numeric}, \code{max_numeric}, and \code{value}.}

\item{round}{\code{TRUE} to round all values to the nearest integer;
\code{FALSE} if no rounding is desired; or an integer to round to that
number of digits (for example, 1 will round to the nearest 10, and -2 will
round to the nearest .01). Any rounding will be applied after snapping to
the nearest step.}

\item{ticks}{\code{FALSE} to hide tick marks, \code{TRUE} to show them
according to some simple heuristics.}

\item{animate}{\code{TRUE} to show simple animation controls with default
settings; \code{FALSE} not to; or a custom settings list, such as those
created using \code{\link[shiny:animationOptions]{animationOptions()}}.}

\item{slider_width}{(scalar) Width of the slider. Must be valid for use as a
CSS unit of length.}

\item{sep}{Separator between thousands places in numbers.}

\item{pre}{A prefix string to put in front of the value.}

\item{post}{A suffix string to put after the value.}
}
\description{
Creates HTML code for a visual analog scale slider.
}
