% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scoring_maps}
\alias{scoring_maps}
\title{Scoring Maps}
\format{
An object of class \code{list} of length 1.
}
\usage{
scoring_maps
}
\description{
This list contains scoring maps (i.e., data frames) for some questionnaires.
}
\keyword{datasets}
