% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-params.R
\name{session_design_url_welcome_page}
\alias{session_design_url_welcome_page}
\title{Welcome page with URL parameter retrieval}
\usage{
session_design_url_welcome_page(
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f",
  debug = FALSE
)
}
\arguments{
\item{dict}{internationalization dictionary.}

\item{default_lang}{Default language for the page.}

\item{debug}{(scalar logical) whether or not to display the URL parameter
\code{udes} (i.e., the design) on the page.}
}
\description{
This functions creates a welcome page.
It checks whether the URL parameters \code{uses} (for session) and \code{udes}
(for design) are present.
After completing the page, both URL parameters are set as global variables
(with \code{\link[psychTestR:global_local]{psychTestR::set_global()}} in the \code{on_complete} argument of the page).
}
