% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/education-profession.R
\name{student_page}
\alias{student_page}
\title{University student page}
\usage{
student_page(
  label = "student",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f"
)
}
\arguments{
\item{label}{(scalar character) Label for the current page.}

\item{dict}{internationalization dictionary.}

\item{default_lang}{(scalar character) default language.}
}
\description{
This function creates a page to ask participants whether they are currently
studying at a university.
}
