% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile-device-landscape.R
\name{use_landscape_page}
\alias{use_landscape_page}
\title{Mobile devices landscape}
\usage{
use_landscape_page(
  img_src = NULL,
  img_width = "100\%",
  img_align = "center",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f"
)
}
\arguments{
\item{img_src}{\code{NULL} for no image and a URL (character scalar) to display
an image.}

\item{img_width}{CSS unit (see \code{\link[shiny:reexports]{shiny::validateCssUnit()}}) for the width
of the image.}

\item{img_align}{(character scalar) sets the alignment of the image.}

\item{dict}{internationalization dictionary.}

\item{default_lang}{Default language for the page.}
}
\description{
This function creates a page prompting participants to use mobile devices
in landscape mode.
}
