% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocaloidproject-package.R
\docType{package}
\name{vocaloidproject-package}
\alias{vocaloidproject}
\alias{vocaloidproject-package}
\title{vocaloidproject: Package for a research project on vocaloid music}
\description{
This package includes material for a research project on vocaloid music.
}
\author{
\strong{Maintainer}: Kilian Sander \email{kilian.sander@hmtm-hannover.de} (\href{https://orcid.org/0000-0003-3770-7483}{ORCID})

}
\keyword{internal}
