% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vocaloidproject_dict_df}
\alias{vocaloidproject_dict_df}
\title{Data frame of the dictionary for the project}
\format{
An object of class \code{data.frame} with 503 rows and 5 columns.
}
\usage{
vocaloidproject_dict_df
}
\description{
This is the project dictionary as a data.frame.
For details on the dictionary, see \link{vocaloidproject_dict}.
}
\keyword{datasets}
