% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BES.R
\name{BES}
\alias{BES}
\title{Basic Empathy Scale (BES)}
\usage{
BES(
  label = "BES",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f"
)
}
\arguments{
\item{label}{(character scalar) label given to the results section.}

\item{dict}{internationalization dictionary.}

\item{default_lang}{(scalar character) default language.}
}
\description{
Basic Empathy Scale (BES)
}
