% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consent_pages.R
\name{consent_page}
\alias{consent_page}
\alias{consent_page_international1}
\title{Consent pages}
\usage{
consent_page(
  dict = vocaloidproject::vocaloidproject_dict,
  consent_text_key = "",
  consent_give_key = "consent_give",
  consent_no_key = "consent_no",
  default_lang = "de_f"
)

consent_page_international1(
  dict = vocaloidproject::vocaloidproject_dict,
  no_consent_back_link = "\%s",
  debug = FALSE
)
}
\arguments{
\item{dict}{internationalization dictionary.}

\item{consent_text_key}{(scalar character) key in \code{dict}.
Information text for informed consent.}

\item{consent_give_key}{(scalar character) key in \code{dict}.
Label for the radio button to provide consent.}

\item{consent_no_key}{(scalar character) key in \code{dict}.
Label for the radio button indicating no consent.}

\item{default_lang}{Default language for the page.}

\item{no_consent_back_link}{link containing the placeholder \verb{\%s} for \code{p_id}
to redirect participants when they do not provide consent.}

\item{debug}{(logical scalar) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} whether or
not to \code{message()} the URL parameter(s).}
}
\description{
These functions generate pages for participants to provide informed consent.
}
\details{
\code{consent_page()} provides a basic interface for generating consent forms
with radio buttons.
\code{consent_page_international1()} uses \code{consent_page()}, sets the correct
defaults for use in \code{\link[=international1]{international1()}} and adds a redirect to Prolific for
participants who do not consent.
}
