% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotional-baseline.R
\name{emotional_baseline_page}
\alias{emotional_baseline_page}
\title{Individual emotional baseline}
\usage{
emotional_baseline_page(
  label = "emotional_baseline",
  dict = vocaloidproject::vocaloidproject_dict,
  vertical = TRUE,
  default_lang = "de_f"
)
}
\arguments{
\item{label}{Label for the current page.}

\item{dict}{internationalization dictionary.}

\item{vertical}{(logical scalar) If \code{TRUE}, radio buttons are created
vertically below each of the items.
If \code{FALSE}, a radio button matrix will be used.}

\item{default_lang}{Default language for the page.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function creates a page examining the emotional baseline of
a participant.
}
