% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first-language.R
\name{first_language_page}
\alias{first_language_page}
\title{First language}
\usage{
first_language_page(
  label = "first_language",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f"
)
}
\arguments{
\item{label}{(character scalar)}

\item{dict}{internationalization dictionary.}

\item{default_lang}{(scalar character) default language.}
}
\description{
First language
}
