% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/international1.R
\name{international1}
\alias{international1}
\title{Battery for International Experiment 1}
\usage{
international1(
  title = "",
  admin_password = "vocaloid",
  researcher_email = "kilian.vogt@hmtm-hannover.de",
  dict = vocaloidproject::vocaloidproject_dict,
  allow_any_p_id_url = TRUE,
  force_p_id_from_url = FALSE,
  languages = c("de_f", "ja"),
  year_range = c(1925, 2007),
  gms_subscales = c("General"),
  logo = NULL,
  technical_error_back_link = "",
  follow_up_link = "",
  language_url_param = "l",
  language_url_codes = c(de = "ger", de_f = "ger", en = "eng", ja = "jpn"),
  debug = FALSE
)
}
\arguments{
\item{title}{The test's title.
This can either be an unnamed character scalar,
in which case the same title will be shown irrespective of internationalisation,
or a named vector of titles with the names
corresponding to language codes.}

\item{admin_password}{Password to access the admin panel.}

\item{researcher_email}{Researcher's email; used in participant help message.}

\item{dict}{internationalization dictionary.}

\item{allow_any_p_id_url}{Whether to allow new participants
to give themselves an arbitrary participant ID by passing it
as a URL parameter.}

\item{force_p_id_from_url}{Whether to force new participants
to provide their participant ID as a URL parameter.}

\item{languages}{Character vector of languages that may be selected
via the URL parameter 'language'. If no language is provided by
the URL parameter, defaults to the first language in this vector.
Languages should be encoded according to ISO 639-2 conventions, lower-case,
e.g. 'en'.}

\item{year_range}{(length 2 int vector) The minimum and maximum year for birth date dropdown boxes.}

\item{gms_subscales}{(character vector) passed to the \code{subscales} argument of
\code{\link[psyquest:GMS]{psyquest::GMS()}}.}

\item{logo}{Path to a logo to display in the header (optional).}

\item{technical_error_back_link}{(character scalar) link for returning to
Prolific when URL parameter \code{uses} or \code{udes} is missing.
It has to contain the placeholder \verb{\%s} for \code{p_id} to redirect
participants correctly.}

\item{follow_up_link}{(character scalar) link for redirecting participants
to the next section of the experiment (on a different platform).}

\item{language_url_param}{(character scalar or \code{NULL}) URL parameter to pass
on the test language of a participant.
If non-\code{NULL}, the URL parameter will be appended to the \code{back_link} and
the language will be read from the participant's session info.}

\item{language_url_codes}{(named character vector or \code{NULL})
A named character vector will be used when \code{language_url_param} is
non-\code{NULL} to translate the psychTestR language codes
(in most cases two lower-case letters according to ISO 639-2 conventions)
to the codes for the redirect link
(e.g., SoSci Survey uses three letter codes like \code{ger} for German, \code{eng} for
English, or \code{jpn} for Japanese).
If \code{NULL}, the character scalar retrieved by
\code{\link[psychTestR:get_session_info]{psychTestR::get_session_info()}}\verb{$language} is used.}

\item{debug}{(logical scalar) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} whether
or not to display debug information.}
}
\description{
Battery for International Experiment 1
}
