% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/international1-monitor.R
\name{international1_monitor}
\alias{international1_monitor}
\title{Monitor App for International Experiment 1}
\usage{
international1_monitor(
  battery_folder_name = "international1-1",
  battery_folder_name2 = NULL,
  external_data = NULL,
  external_data_type = "sosci",
  title = "Vocaloid Project Experiment 4 (International 1) Monitor",
  data_pw = "supersecretpassword"
)
}
\arguments{
\item{battery_folder_name}{(character scalar) This is the name of the folder
on your shiny server where the \code{app.R} containing \code{international1(...)}
lives.}

\item{battery_folder_name2}{(\code{NULL} or character scalar) When there are two
instances of the \code{international1} battery (e.g., one for each of the two
sessions), specify the second instance here.
If there is only one, set \code{battery_folder_name2} to \code{NULL}.}

\item{external_data}{(\code{NULL} or character scalar) if non-\code{NULL}, a URL to
load extra data from.}

\item{external_data_type}{(character scalar) specifies the import script for
data to load from \code{external_data}.
Currently, the only option is \code{sosci} for the SoSci Survey API for data
retrieval as a csv for R.}

\item{title}{(character scalar) Title for the app.}

\item{data_pw}{(character scalar) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
set a password for data access.
Caution: This is a very simple implementation.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function calls a simple monitor app for International Experiment 1 as
defined in \code{\link[=international1]{international1()}}.
The function assumes that it is running on a shiny server.
}
