% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_page.R
\name{last_page_redirect}
\alias{last_page_redirect}
\title{Last page of psychTestR questionnaire}
\usage{
last_page_redirect(
  redirect_heading = "thanks",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f",
  back_link,
  back_link_key = "return_to_prolific"
)
}
\arguments{
\item{dict}{internationalization dictionary.}

\item{default_lang}{(scalar character) default language.}

\item{back_link}{link containing the placeholder \verb{\%s} for \code{p_id} to redirect
participants.}
}
\description{
Generate the last page for a \code{psychTestR} questionnaire.
}
