% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirect_page.R
\name{last_page_redirect}
\alias{last_page_redirect}
\alias{last_page_redirect_session_design}
\title{Last page of psychTestR questionnaire}
\usage{
last_page_redirect(
  redirect_heading = "thanks",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f",
  back_link,
  back_link_key = "return_to_prolific",
  debug = FALSE
)

last_page_redirect_session_design(
  redirect_heading = "thanks",
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f",
  back_link,
  back_link_key = "CONTINUE",
  debug = FALSE
)
}
\arguments{
\item{redirect_heading}{(scalar character or \code{NULL}) Key in \code{dict}.
Its translation is shown as the heading of this page.
\code{NULL} suppresses the heading.}

\item{dict}{internationalization dictionary.}

\item{default_lang}{(scalar character) default language.}

\item{back_link}{link containing the placeholder \verb{\%s} for \code{p_id} to redirect
participants.
In the case of \code{last_page_redirect_session_design} two additional
placeholders for session and design have to be included.}

\item{back_link_key}{(scalar character) Key in \code{dict}. Its translation is
the text shown for \code{back_link}.}

\item{debug}{(logical scalar) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} whether or
not to \code{message()} the URL parameter(s).}
}
\description{
Generate the last page for a \code{psychTestR} questionnaire including a redirect
link.
}
\details{
\code{p_id} is passed on as a URL parameter.
\code{last_page_redirect_session_design()} also passes on \code{uses} (session)
and \code{udes} (design).
}
