% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redirect_page.R
\name{last_page_redirect}
\alias{last_page_redirect}
\alias{last_page_redirect_session_design}
\title{Last page of psychTestR questionnaire}
\usage{
last_page_redirect(
  redirect_heading = "thanks",
  redirect_paragraph = NULL,
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f",
  back_link,
  back_link_key = "return_to_prolific",
  debug = FALSE
)

last_page_redirect_session_design(
  redirect_heading = "thanks",
  redirect_paragraph = NULL,
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f",
  back_link,
  back_link_key = "CONTINUE",
  language_url_param = "l",
  language_url_codes = c(de = "ger", de_f = "ger", en = "eng", ja = "jpn"),
  debug = FALSE
)
}
\arguments{
\item{redirect_heading}{(scalar character or \code{NULL}) Key in \code{dict}.
Its translation is displayed as the heading of this page.
\code{NULL} suppresses the heading.}

\item{redirect_paragraph}{(scalar character or \code{NULL}) Key in \code{dict}.
Its translation is displayed as a paragraph of text on the page.
\code{NULL} suppresses a paragraph.}

\item{dict}{internationalization dictionary.}

\item{default_lang}{(scalar character) default language.}

\item{back_link}{link containing the placeholder \verb{\%s} for \code{p_id} to redirect
participants.
In the case of \code{last_page_redirect_session_design} two additional
placeholders for session and design have to be included.}

\item{back_link_key}{(scalar character) Key in \code{dict}. Its translation is
the text shown for \code{back_link}.}

\item{debug}{(logical scalar) \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} whether or
not to \code{message()} the URL parameter(s).}

\item{language_url_param}{(character scalar or \code{NULL}) URL parameter to pass
on the test language of a participant.
If non-\code{NULL}, the URL parameter will be appended to the \code{back_link} and
the language will be read from the participant's session info.}

\item{language_url_codes}{(named character vector or \code{NULL})
A named character vector will be used when \code{language_url_param} is
non-\code{NULL} to translate the psychTestR language codes
(in most cases two lower-case letters according to ISO 639-2 conventions)
to the codes for the redirect link
(e.g., SoSci Survey uses three letter codes like \code{ger} for German, \code{eng} for
English, or \code{jpn} for Japanese).
If \code{NULL}, the character scalar retrieved by
\code{\link[psychTestR:get_session_info]{psychTestR::get_session_info()}}\verb{$language} is used.}
}
\description{
Generate the last page for a \code{psychTestR} questionnaire including a redirect
link.
}
\details{
\code{p_id} is passed on as a URL parameter.
\code{last_page_redirect_session_design()} also passes on \code{uses} (session)
and \code{udes} (design).
Optionally, you can specify a language URL parameter to pass on
the participant's test language.
}
