% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url-params.R
\name{design_url_welcome_page}
\alias{design_url_welcome_page}
\alias{session_design_url_welcome_page}
\title{Welcome page with URL parameter retrieval}
\usage{
design_url_welcome_page(
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f",
  session_number = 1,
  debug = FALSE
)

session_design_url_welcome_page(
  dict = vocaloidproject::vocaloidproject_dict,
  default_lang = "de_f",
  debug = FALSE
)
}
\arguments{
\item{dict}{internationalization dictionary.}

\item{default_lang}{Default language for the page.}

\item{session_number}{(integer-like scalar) sets the session number.}

\item{debug}{(logical scalar) whether or not to display the URL parameter
\code{udes} (i.e., the design) on the page.}
}
\description{
These functions create a welcome page.
They check whether the URL parameter \code{udes} (for design) is present and
in the case of \code{session_design_url_welcome_page()}, whether \code{uses}
(for session) is also present.
After completing the page, each URL parameter is set as a global variable
(with \code{\link[psychTestR:global_local]{psychTestR::set_global()}} in the \code{on_complete} argument of the page).
}
