% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocaloid_battery.R
\name{vocaloid_battery}
\alias{vocaloid_battery}
\title{Battery}
\usage{
vocaloid_battery(
  dict = vocaloidproject::vocaloidproject_dict,
  languages = c("de_f", "ja"),
  back_link
)
}
\arguments{
\item{dict}{internationalization dictionary.}

\item{languages}{(character vector) languages for the battery of questionnaires}

\item{back_link}{link to redirect participants.}
}
\description{
Battery
}
