% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocaloid_battery.R
\name{vocaloid_battery}
\alias{vocaloid_battery}
\title{Battery}
\usage{
vocaloid_battery(
  title = "vocaloidproject",
  admin_password = "vocaloid",
  researcher_email = NULL,
  demo = FALSE,
  dict = vocaloidproject::vocaloidproject_dict,
  allow_any_p_id_url = TRUE,
  force_p_id_from_url = FALSE,
  languages = c("de_f", "ja"),
  back_link = NA_character_,
  gms_subscales = c("Abilities"),
  logo = NULL
)
}
\arguments{
\item{dict}{internationalization dictionary.}

\item{languages}{(character vector) languages for the battery of questionnaires}

\item{back_link}{link to redirect participants.}
}
\description{
Battery
}
