% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice-rating-battery.R
\name{voice_rating_battery}
\alias{voice_rating_battery}
\title{psychTestR Battery for Voice Rating Study}
\usage{
voice_rating_battery(
  title = "Voiceratingbattery",
  admin_password = "vocaloid",
  randomise_stimuli_at_runtime = TRUE,
  researcher_email = NULL,
  demo = FALSE,
  dict = vocaloidproject::vocaloidproject_dict,
  allow_any_p_id_url = TRUE,
  force_p_id_from_url = FALSE,
  languages = c("de_f"),
  stimulus_url = character(),
  logo = NULL,
  debug = FALSE
)
}
\description{
psychTestR Battery for Voice Rating Study
}
